moves the hook to the relevant folder
